Promote script to command
